requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.7
eventlet==0.36.1
Werkzeug==3.0.4
python-dotenv==1.0.1
